+++ geoip2/models.py
@@ -17,10 +17,10 @@ from abc import ABCMeta
 from typing import Any, cast, Dict, Optional, Sequence, Union
 
 import geoip2.records
-from geoip2.mixins import SimpleEquality
+from geoip2.mixins import Model
 
 
-class Country(SimpleEquality):
+class Country(Model):
     """Model for the Country web service and Country database.
 
     This class provides the following attributes:
@@ -97,12 +97,9 @@ class Country(SimpleEquality):
         self.maxmind = geoip2.records.MaxMind(**raw_response.get("maxmind", {}))
 
         self.traits = geoip2.records.Traits(**raw_response.get("traits", {}))
-        self.raw = raw_response
 
     def __repr__(self) -> str:
-        return (
-            f"{self.__module__}.{self.__class__.__name__}({self.raw}, {self._locales})"
-        )
+        return f"{self.__module__}.{self.__class__.__name__}({self.to_dict()}, {self._locales})"
 
 
 class City(Country):
@@ -321,22 +318,27 @@ class Enterprise(City):
     """
 
 
-class SimpleModel(SimpleEquality, metaclass=ABCMeta):
+class SimpleModel(Model, metaclass=ABCMeta):
     """Provides basic methods for non-location models"""
 
-    raw: Dict[str, Union[bool, str, int]]
     ip_address: str
     _network: Optional[Union[ipaddress.IPv4Network, ipaddress.IPv6Network]]
     _prefix_len: int
 
     def __init__(self, raw: Dict[str, Union[bool, str, int]]) -> None:
-        self.raw = raw
-        self._network = None
-        self._prefix_len = cast(int, raw.get("prefix_len"))
+        if network := raw.get("network"):
+            self._network = ipaddress.ip_network(network, False)
+            self._prefix_len = self._network.prefixlen
+        else:
+            # This case is for MMDB lookups where performance is paramount.
+            # This is why we don't generate the network unless .network is
+            # used.
+            self._network = None
+            self._prefix_len = cast(int, raw.get("prefix_len"))
         self.ip_address = cast(str, raw.get("ip_address"))
 
     def __repr__(self) -> str:
-        return f"{self.__module__}.{self.__class__.__name__}({self.raw})"
+        return f"{self.__module__}.{self.__class__.__name__}({self.to_dict()})"
 
     @property
     def network(self) -> Optional[Union[ipaddress.IPv4Network, ipaddress.IPv6Network]]: