+++ HISTORY.rst
@@ -4,9 +4,12 @@
 History
 -------
 
-4.9.0
+5.0.0
 ++++++++++++++++++
 
+* BREAKING: The ``raw`` attribute on the model classes has been replaced
+  with a ``to_dict()`` method. This can be used to get a representation of
+  the object that is suitable for serialization.
 * IMPORTANT: Python 3.9 or greater is required. If you are using an older
   version, please use an earlier release.
 * ``metro_code`` on ``geoip2.record.Location`` has been deprecated. The